Use JSX syntax and extract getSourceIcon to utils
