+++ mlflow/server/js/src/shared/web-shared/model-trace-explorer/assessments-pane/AssessmentItemHeader.tsx
@@ -7,17 +7,7 @@ import { AssessmentDeleteModal } from './AssessmentDeleteModal';
 import { AssessmentSourceName } from './AssessmentSourceName';
 import { timeSinceStr } from './AssessmentsPane.utils';
 import type { Assessment } from '../ModelTrace.types';
-
-export const getSourceIcon = (source: Assessment['source']) => {
-  switch (source.source_type) {
-    case 'HUMAN':
-      return UserIcon;
-    case 'LLM_JUDGE':
-      return SparkleIcon;
-    default:
-      return CodeIcon;
-  }
-};
+import { getSourceIcon } from './utils';
 
 export const AssessmentItemHeader = ({
   // connector is not displayed in history items
