+++ mlflow/server/js/src/shared/web-shared/model-trace-explorer/assessments-pane/utils.tsx
@@ -1,4 +1,5 @@
 import type { Assessment } from '../ModelTrace.types';
+import { useDesignSystemTheme, Typography, SparkleIcon, UserIcon, CodeIcon } from '@databricks/design-system';
 
 export const getAssessmentValue = (assessment: Assessment) => {
   if ('feedback' in assessment) {
@@ -10,4 +11,15 @@ export const getAssessmentValue = (assessment: Assessment) => {
   }
 
   return assessment.expectation.serialized_value.value;
+};
+
+export const getSourceIcon = (source: Assessment['source']) => {
+  switch (source.source_type) {
+    case 'HUMAN':
+      return UserIcon;
+    case 'LLM_JUDGE':
+      return SparkleIcon;
+    default:
+      return CodeIcon;
+  }
 };
